Refactor to interface
