+++ platforms/platform-alexa/src/interfaces.ts
@@ -222,3 +222,35 @@ export interface Request {
 
 // Defines a target for Alexa Conversations
 export type ConversationsTarget = 'AMAZON.Conversations' | 'skill';
+
+export interface ConnectionPostalAddress {
+  "@type": "PostalAddress",
+  "@version": "1",
+  streetAddress: string;
+  locality: string;
+  region: string;
+  postalCode: string;
+  country?: string;
+}
+
+export enum ConsentLevel {
+  Account = 'ACCOUNT',
+  Person = 'PERSON',
+}
+
+export type ConsentLevelLike = EnumLike<ConsentLevel> | string;
+
+export enum PermissionScope {
+  ReadProfileGivenName = 'alexa::profile:given_name:read',
+  ReadProfileName = 'alexa::profile:name:read',
+  ReadProfileMobileNumber = 'alexa::profile:mobile_number:read',
+  ReadProfileEmail = 'alexa::profile:email:read',
+  ReadAddressCountryAndPostalCode = 'alexa:devices:all:address:country_and_postal_code:read',
+  ReadGeolocation = 'alexa::devices:all:geolocation:read',
+  ReadWriteTimers = 'alexa::alerts:timers:skill:readwrite',
+  ReadWriteReminders = 'alexa::alerts:reminders:skill:readwrite',
+}
+
+export type PermissionScopeLike = EnumLike<PermissionScope> | string;
+
+
