+++ platforms/platform-alexa/src/output/templates/ConnectionAskForPermissionConsentOutput.ts
@@ -1,24 +1,5 @@
-import { BaseOutput, Output, OutputOptions, OutputTemplate, EnumLike } from '@jovotech/framework';
-
-export enum ConsentLevel {
-  Account = 'ACCOUNT',
-  Person = 'PERSON',
-}
-
-export type ConsentLevelLike = EnumLike<ConsentLevel> | string;
-
-export enum PermissionScope {
-  ReadProfileGivenName = 'alexa::profile:given_name:read',
-  ReadProfileName = 'alexa::profile:name:read',
-  ReadProfileMobileNumber = 'alexa::profile:mobile_number:read',
-  ReadProfileEmail = 'alexa::profile:email:read',
-  ReadAddressCountryAndPostalCode = 'alexa:devices:all:address:country_and_postal_code:read',
-  ReadGeolocation = 'alexa::devices:all:geolocation:read',
-  ReadWriteTimers = 'alexa::alerts:timers:skill:readwrite',
-  ReadWriteReminders = 'alexa::alerts:reminders:skill:readwrite',
-}
-
-export type PermissionScopeLike = EnumLike<PermissionScope> | string;
+import { BaseOutput, Output, OutputOptions, OutputTemplate } from '@jovotech/framework';
+import { PermissionScopeLike, ConsentLevelLike } from '../../interfaces';
 
 export interface PermissionScopeItem {
   permissionScope: PermissionScopeLike;
