--- platforms/platform-alexa/src/output/templates/ConnectionPostalAddress.ts
@@ -1,9 +1,0 @@
-export interface ConnectionPostalAddress {
-  "@type": "PostalAddress",
-  "@version": "1",
-  streetAddress: string;
-  locality: string;
-  region: string;
-  postalCode: string;
-  country?: string;
-}
